AST: Move AvailabilityContext into its own header and cpp file.
