+++ include/swift/AST/TypeRefinementContext.h
@@ -19,6 +19,7 @@
 #define SWIFT_TYPEREFINEMENTCONTEXT_H
 
 #include "swift/AST/Availability.h"
+#include "swift/AST/AvailabilityContext.h"
 #include "swift/AST/Identifier.h"
 #include "swift/AST/Stmt.h" // for PoundAvailableInfo
 #include "swift/Basic/Debug.h"
